requirements.txt
pandas
numpy
psycopg[binary]
python-dotenv
matplotlib
seaborn
scikit-learn
statsmodels
SQLAlchemy
jupyter
